feat(new-marquee): extracted into its own block
