+++ express/blocks/columns/columns.css
@@ -769,223 +769,6 @@ main .collapsible-card-wrapper + .columns-wrapper .columns > div {
   }
 }
 
-main .columns-fullscreen-center-container {
-  min-height: 460px;
-  box-sizing: border-box;
-  padding: 160px 0;
-  position: relative;
-  height: 100%;
-}
-
-main .columns-fullscreen-center-container .marquee-background {
-  content: '';
-  display: block;
-  position: absolute;
-  left: 0;
-  top: 0;
-  min-width: 100%;
-  height: 100%;
-  background-image: url('https://www.adobe.com/express/media_1aa259dd9ac4c29168d486d36c6c737074d24d756.jpeg#width=1600&height=901');
-  background-repeat: no-repeat;
-  background-size: cover;
-  opacity: 0.1;
-  transition: opacity 0.1s;
-}
-
-main .columns-fullscreen-center-container .columns-marquee h1 {
-  font-size: 80px;
-  line-height: 90px;
-}
-
-main .columns-fullscreen-center-container .columns-marquee p {
-  max-width: 650px;
-  margin: 16px auto;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .button-container > a.button {
-  width: 180px;
-  padding: 12px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 120px;
-}
-
-main .columns-fullscreen-center-container .columns-marquee h1 > span.colorful {
-  background: linear-gradient(320deg, #7C84F3, #FF4DD2, #FF993B, #FF4DD2, #7C84F3, #FF4DD2, #FF993B);
-  background-size: 400% 400%;
-  -webkit-animation: buttonGradient 45s ease infinite;
-  -moz-animation: buttonGradient 45s ease infinite;
-  animation: buttonGradient 45s ease infinite;
-  -webkit-background-clip: text;
-  background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .has-brand .column:not(.column-picture):not(.hero-animation-overlay) {
-  padding: 24px;
-}
-
-main .section.columns-fullscreen-center-container div.columns-marquee > div > div.picture-frame {
-  position: relative;
-  background-image: url('https://www.adobe.com/express/media_13c6a255a99e9ab7a39bdc21ca9b86652c3e59f27.png#width=1600&height=1077');
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: contain;
-  height: 100%;
-  width: 100%;
-  border-radius: 24px;
-  display: block;
-  z-index: 1;
-  transform-style: preserve-3d;
-  transition: transform 0.5s ease-out;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .picture-frame:before {
-  content: '';
-  position: absolute;
-  border-radius: 24px;
-  height: 100%;
-  width: 100%;
-  background-color: #FFFFFF;
-  z-index: -1;
-  transform: translateZ(26px);
-  transition: opacity 0.2s;
-  opacity: 0;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper {
-  padding: 16px;
-  height: 518px;
-  width: 770px;
-  margin: auto;
-  will-change: transform;
-  border-radius: 24px;
-  transition: transform 0.5s ease-out;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-background {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  border-radius: 42px;
-  background-color: #f3f3f3;
-  box-shadow: 0 0 40px 20px hsl(0deg 0% 63% / 24%);
-  transition: transform 0.5s ease-out;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .screen-demo {
-  position: absolute;
-  top: 64px;
-  right: 72px;
-  height: 84%;
-  z-index: 2;
-  transform-style: preserve-3d;
-  transform: translate3d(0px, 0px, 15px);
-  transition: transform 0.5s ease-out, opacity 0.2s;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .picture-frame-clickable-layer {
-  position: absolute;
-  height: 100%;
-  width: 100%;
-  cursor: pointer;
-  z-index: 3;
-  transform: translateZ(100px);
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .picture-frame-clickable-layer > .button {
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  padding: 12px 32px;
-  border-radius: 32px;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .picture-frame-clickable-layer:hover ~ .flowers .flower,
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .picture-frame-clickable-layer:hover ~ .picture-frame:before {
-  opacity: 0.5;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .flowers {
-  position: absolute;
-  display: block;
-  z-index: -1;
-  height: 100%;
-  width: 100%;
-  transform-style: preserve-3d;
-  backface-visibility: hidden;
-  transition: transform 0.5s ease-out;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .flowers .flower {
-  position: absolute;
-  transition: opacity 0.5s ease-out;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .flower.flower-0 {
-  top: 40px;
-  left: -80px;
-  width: 115px;
-  transform: rotate(-50deg);
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .flower.flower-1 {
-  top: 130px;
-  left: -156px;
-  width: 223px;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .flower.flower-2 {
-  bottom: 72px;
-  left: -91px;
-  width: 111px;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .flower.flower-3 {
-  top: 110px;
-  right: -100px;
-  width: 142px;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .flower.flower-4 {
-  top: 180px;
-  right: -150px;
-  width: 181px;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-wrapper .flower.flower-5 {
-  bottom: 40px;
-  right: -115px;
-  width: 154px;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-thumbnails {
-  position: absolute;
-  display: block;
-  left: 68px;
-  bottom: 8px;
-  height: 70%;
-  width: 180px;
-  transform-style: preserve-3d;
-  backface-visibility: hidden;
-  transition: transform 0.5s ease-out;
-  transform: translate3d(0px, 0px, 15px);
-  z-index: 1;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame-thumbnails .leaf-thumbnails {
-  display: block;
-  height: 100%;
-  width: 100%;
-}
-
-main .columns-fullscreen-center-container .columns-marquee .picture-frame picture img {
-  height: 100%;
-  display: block;
-}
-
 /* Japanese font sizing styles */
 :lang(ja) main .columns h2.columns-heading-long {
   font-size: var(--heading-font-size-l);
